Group14/Astraeus-Multi-Agent-AI-Researcher-main/requirements.txt
streamlit>=1.28.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
plotly>=5.18.0
numpy>=1.24.0
pandas>=1.5.0
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
pypdf>=4.0.0
